requirements.txt
cryptography
django
django-fernet-encrypted-fields
djangorestframework
django-filter
inflection
uritemplate
whitenoise
django-environ